chore: Start adding `AutomationPeer` implementation
