+++ src/Uno.UI/UI/Xaml/Automation/Peers/AutomationPeer.mux.cs
@@ -4,6 +4,31 @@ namespace Microsoft.UI.Xaml.Automation.Peers
 {
 	public partial class AutomationPeer
 	{
+		private bool HasKeyboardFocusHelper()
+		{
+			xref_ptr<CAutomationPeer> ap;
+			XINT32 isKeyboardFocusable = FALSE;
+
+			*pRetVal = FALSE;
+
+			IFC_RETURN(IsKeyboardFocusable(&isKeyboardFocusable));
+			if (isKeyboardFocusable)
+			{
+				if (CFocusManager * focusManager = GetFocusManagerNoRef())
+				{
+					ap = focusManager->GetFocusedAutomationPeer();
+
+					if (this == static_cast<CAutomationPeer*>(ap.get()))
+					{
+						*pRetVal = TRUE;
+					}
+				}
+			}
+
+			return S_OK;
+		}
+
+
 		internal static bool ListenerExistsHelper(AutomationEvents eventId)
 #if __SKIA__
 			=> AutomationPeerListener?.ListenerExistsHelper(eventId) == true;
